RGPS is finally working

+++ src/PlotterRel.hpp
@@ -12,7 +12,7 @@
 
 class PlotterRel {
 public:
-    PlotterRel(const SatNavRel& sn, unsigned ti = 0, unsigned tf = 0);
+    PlotterRel(const SatNavRel& sn, double ti = 0, double tf = 0);
     void plot_errors_norm(double ymin = 0, double ymax = 0);
 
 private:
